// api/model/connectorStatus.ts
/**
 * CPMS API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export type ConnectorStatus = 'Undefined' | 'Available' | 'Preparing' | 'Charging' | 'SuspendedEV' | 'SuspendedEVSE' | 'Finishing' | 'Reserved' | 'Unavailable' | 'Faulted';

export const ConnectorStatus = {
    Undefined: 'Undefined' as ConnectorStatus,
    Available: 'Available' as ConnectorStatus,
    Preparing: 'Preparing' as ConnectorStatus,
    Charging: 'Charging' as ConnectorStatus,
    SuspendedEV: 'SuspendedEV' as ConnectorStatus,
    SuspendedEVSE: 'SuspendedEVSE' as ConnectorStatus,
    Finishing: 'Finishing' as ConnectorStatus,
    Reserved: 'Reserved' as ConnectorStatus,
    Unavailable: 'Unavailable' as ConnectorStatus,
    Faulted: 'Faulted' as ConnectorStatus
};