// api/model/connectorType.ts
/**
 * AdminApp API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export type ConnectorType = 'Undefined' | 'Type1Connector' | 'Type2Connector' | 'CHAdeMO' | 'CCS1Combo' | 'CCS2Combo' | 'TeslaConnector' | 'GB_TConnector' | 'Other';

export const ConnectorType = {
    Undefined: 'Undefined' as ConnectorType,
    Type1Connector: 'Type1Connector' as ConnectorType,
    Type2Connector: 'Type2Connector' as ConnectorType,
    CHAdeMO: 'CHAdeMO' as ConnectorType,
    CCS1Combo: 'CCS1Combo' as ConnectorType,
    CCS2Combo: 'CCS2Combo' as ConnectorType,
    TeslaConnector: 'TeslaConnector' as ConnectorType,
    GB_TConnector: 'GB_TConnector' as ConnectorType,
    Other: 'Other' as ConnectorType
};