// api/model/connectorExtendedDto.ts
/**
 * CPMS API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { ConnectorType } from './connectorType';
import { ConnectorStatus } from './connectorStatus';
import { VoltageLevel } from './voltageLevel';
import { ChargingPowerType } from './chargingPowerType';
import { ChargePointStatus } from './chargePointStatus';
import { TariffSimpleDto } from './tariffSimpleDto';
import { PowerType } from './powerType';
import { FormatType } from './formatType';


export interface ConnectorExtendedDto { 
    id?: string | null;
    chargePointId?: string | null;
    powerType?: PowerType;
    formatType?: FormatType;
    status?: ConnectorStatus;
    connectorType?: ConnectorType;
    connectorNumber?: number;
    voltageLevel?: VoltageLevel;
    currentType?: ChargingPowerType;
    externalId?: string | null;
    amps?: number | null;
    watts?: number | null;
    lastStatusTime?: string | null;
    lastMeter?: number | null;
    lastMeterTime?: string | null;
    tariff?: TariffSimpleDto;
    currentSoC?: number | null;
    qrCodeValue?: string | null;
    ocppInfo?: string | null;
    ocppVendorId?: string | null;
    ocppErrorCode?: string | null;
    ocppVendorErrorCode?: string | null;
    chargePointReadableId?: string | null;
    chargePointName?: string | null;
    chargePointStatus?: ChargePointStatus;
    workspaceId?: string | null;
    chargePointLastHeartBeatTime?: string | null;
}
export namespace ConnectorExtendedDto {
}